Our PV-on-the-stack implementation doesn't need to be used for move ordering anymore, since the hash table usually does a good enough job.

+++ src/board.h
@@ -83,8 +83,8 @@ typedef struct board_s
 	uint8 trades;
 
 	// move ordering:
-	uint32 victim [33];
-	uint32 attack [33];
+	uint16 victim [33];
+	uint16 attack [33];
 } board;
 
 extern board *curboard;
